Organized internal session type
