--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionApproveParams.swift
@@ -1,11 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    struct ApproveParams: Codable, Equatable {
-        let relay: RelayProtocolOptions
-        let responder: SessionParticipant
-        let expiry: Int
-        let state: SessionState
-    }
-}
