--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionPingParams.swift
@@ -1,6 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    struct PingParams: Codable, Equatable {}
-}
