--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionProposeParams.swift
@@ -1,6 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    typealias ProposeParams = SessionProposal
-}
