--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionRejectParams.swift
@@ -1,8 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    struct RejectParams: Codable, Equatable {
-        let reason: Reason
-    }
-}
