--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionUpdateParams.swift
@@ -1,8 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    struct UpdateParams: Codable, Equatable {
-        let state: SessionState
-    }
-}
