--- Sources/WalletConnect/Types/Session/ClientSynchronisation/SessionUpgradeParams.swift
@@ -1,8 +1,0 @@
-
-import Foundation
-
-extension SessionType {
-    struct UpgradeParams: Codable, Equatable {
-        let permissions: SessionPermissions
-    }
-}
